+++ src/descriptors/_element_center_test.js
@@ -22,10 +22,10 @@ describe("ElementCenter", function() {
 	beforeEach(function() {
 		var frame = reset.frame;
 
-		frame.addElement(
+		frame.add(
 			"<p id='one' style='position: absolute; left: 20px; width: 130px; top: 60px; height: 60px'>one</p>"
 		);
-		element = frame.getElement("#one");
+		element = frame.get("#one");
 		center = ElementCenter.x(element);
 		middle = ElementCenter.y(element);
 	});