+++ src/descriptors/_element_size_test.js
@@ -19,10 +19,10 @@ describe("ElementSize", function() {
 
 	beforeEach(function() {
 		var frame = reset.frame;
-		frame.addElement(
+		frame.add(
 			"<p id='element' style='position: absolute; left: 20px; width: 130px; top: 10px; height: 60px'>element</p>"
 		);
-		element = frame.getElement("#element");
+		element = frame.get("#element");
 		width = ElementSize.x(element);
 		height = ElementSize.y(element);
 	});