+++ src/descriptors/_relative_size_test.js
@@ -21,10 +21,10 @@ describe("RelativeSize", function() {
 
 	beforeEach(function() {
 		var frame = reset.frame;
-		frame.addElement(
+		frame.add(
 			"<p id='element' style='position: absolute; left: 20px; width: 130px; top: 10px; height: 60px'>element</p>"
 		);
-		element = frame.getElement("#element");
+		element = frame.get("#element");
 		larger = RelativeSize.larger(element.height, 10);
 		smaller = RelativeSize.smaller(element.width, 5);
 	});