+++ src/descriptors/_size_multiple_test.js
@@ -18,10 +18,10 @@ describe("SizeMultiple", function() {
 
 	beforeEach(function() {
 		var frame = reset.frame;
-		frame.addElement(
+		frame.add(
 			"<p id='element' style='position: absolute; left: 20px; width: 130px; top: 10px; height: 60px'>element</p>"
 		);
-		element = frame.getElement("#element");
+		element = frame.get("#element");
 		twice = SizeMultiple.create(element.width, 2);
 	});